chat: Moves `Color` to `types::consts` and renames the reader method.
